+++ sandpack-react/src/contexts/themeContext.tsx
@@ -1,7 +1,7 @@
 import { useClasser } from "@code-hike/classer";
 import * as React from "react";
 
-import { createThemeObject, codesandboxLightTheme } from "../themes";
+import { createThemeObject, defaultLight } from "../themes";
 import type { SandpackTheme, SandpackThemeProp } from "../types";
 import { injectThemeStyleSheet } from "../utils/domUtils";
 
@@ -12,8 +12,8 @@ const SandpackThemeContext = React.createContext<{
   theme: SandpackTheme;
   id: string;
 }>({
-  theme: codesandboxLightTheme,
-  id: "codesandbox-light",
+  theme: defaultLight,
+  id: "light",
 });
 
 const SandpackThemeProvider: React.FC<{