+++ website/docs/docusaurus.config.js
@@ -128,6 +128,8 @@ module.exports = {
         docs: {
           routeBasePath: "/",
           sidebarPath: require.resolve("./sidebars.js"),
+          sidebarCollapsed: false,
+
           editUrl:
             "https://github.com/codesandbox/sandpack/edit/main/website/docs",
         },
@@ -149,7 +151,7 @@ module.exports = {
   themeConfig:
     /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
     ({
-      sandpack: { theme: require("./src/scss/sandpack-theme") },
+      sandpack: { theme: "sandpack-dark" },
       colorMode: {
         defaultMode: "dark",
         disableSwitch: true,
@@ -161,14 +163,6 @@ module.exports = {
           src: "img/logo-dark.svg",
           srcDark: "img/logo.svg",
         },
-        items: [
-          { to: "/releases", label: "Releases", position: "right" },
-          {
-            href: "https://github.com/codesandbox/sandpack",
-            label: "GitHub",
-            position: "right",
-          },
-        ],
       },
       footer: {
         style: "dark",