+++ website/landing/components/AdvancedUsage/illustrations/index.ts
@@ -1,0 +1,3 @@
+export { ClientIllustration } from "./Client/Client";
+export { ComponentsIllustration } from "./Components/Components";
+export { ProviderIllustration } from "./Provider/Provider";