+++ website/landing/components/common/index.ts
@@ -1,9 +1,11 @@
 export { Box } from "./Box";
 export { Button } from "./Button";
+export * from "./Card";
 export { CodeBlock } from "./CodeBlock";
 export { Clipboard } from "./Clipboard";
-export { List } from "./List";
+export * from "./List";
 export { Resources } from "./Resources";
+export * from "./Section";
 export { SandpackLogo } from "./SandpackLogo";
 export { SandpackPreview } from "./SandpackPreview";
 export { Text } from "./Text";