+++ website/landing/pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 
-import { Api } from "../components/Api";
+import { AdvancedUsage } from "../components/AdvancedUsage";
 import { Banner } from "../components/Banner";
 import { Community } from "../components/Community";
 import { Features } from "../components/Features";
@@ -12,6 +12,8 @@ import { Intro } from "../components/Intro";
 import { Users } from "../components/Users";
 import { styled } from "../stitches.config";
 import content from "../website.config.json";
+
+const DEFAULT_HOST = "";
 
 const Container = styled("section", {
   display: "flex",
@@ -27,8 +29,14 @@ const Main = styled("main", {
   flex: 1,
 });
 
-const Home: NextPage = () => {
-  const { global } = content;
+interface HomeProps {
+  host: string | undefined;
+}
+const Home: NextPage<HomeProps> = ({ host }) => {
+  const { global, meta } = content;
+  const HOST_URL = host
+    ? `${host.includes("localhost") ? "http" : "https"}://${host}`
+    : DEFAULT_HOST;
 
   return (
     <Container>
@@ -36,13 +44,24 @@ const Home: NextPage = () => {
         <title>{global.title}</title>
         <meta content={global.description} name="description" />
         <link href="/favicon.ico" rel="icon" />
+
+        {/* Open Graph */}
+        {meta.map(({ name, value }) => {
+          const content =
+            {
+              "og:url": HOST_URL,
+              "og:image": `${HOST_URL}/${value}`,
+            }[name] || value;
+
+          return <meta key={name} content={content} name={name} />;
+        })}
       </Head>
 
       <Main>
         <Hero />
         <Intro />
         <Features />
-        <Api />
+        <AdvancedUsage />
         {/* <Showcase />*/}
         <Users />
         <Banner />
@@ -51,6 +70,13 @@ const Home: NextPage = () => {
       <Footer />
     </Container>
   );
+};
+
+Home.getInitialProps = async (context) => {
+  const { req } = context;
+  const host = req?.headers.host;
+
+  return { host };
 };
 
 export default Home;