+++ website/docs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/codesandbox/sandpack/compare/v0.5.4...v0.6.0) (2021-11-22)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.5.4](https://github.com/codesandbox/sandpack/compare/v0.5.3...v0.5.4) (2021-11-19)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.5.3](https://github.com/codesandbox/sandpack/compare/v0.5.2...v0.5.3) (2021-11-19)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+# [0.5.0](https://github.com/codesandbox/sandpack/compare/v0.4.1...v0.5.0) (2021-11-19)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.4.1](https://github.com/codesandbox/sandpack/compare/v0.4.0...v0.4.1) (2021-11-18)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
 # [0.4.0](https://github.com/codesandbox/sandpack/compare/v0.3.10...v0.4.0) (2021-11-18)
 
 **Note:** Version bump only for package sandpack-docs